pyproject.toml
 [build-system]
 requires = ["setuptools>=68", "wheel"]
 build-backend = "setuptools.build_meta"

 [project]
 name = "arbitrage-bot"
 version = "0.1.0"
 description = "Modular multi-exchange crypto arbitrage bot"
 readme = "README.md"
 requires-python = ">=3.11"
 authors = [
   { name = "Arbitrage Team" }
 ]
 license = { text = "MIT" }
 dependencies = [
   "aiohttp>=3.9.4",
   "websockets>=12.0",
   "pydantic>=2.5",
   "pydantic-settings>=2.2",
   "structlog>=24.1",
   "PyYAML>=6.0",
   "Flask>=3.0",
   "Flask-SocketIO>=5.3",
   "python-socketio[client]>=5.11",
   "aiogram>=3.4",
   "python-dotenv>=1.0",
   "redis>=5.0",
   "uvicorn>=0.30",
   "orjson>=3.10"
 ]

 [project.optional-dependencies]
 dev = [
   "pytest>=7.4",
   "pytest-asyncio>=0.23",
   "pytest-cov>=4.1",
   "mypy>=1.10",
   "ruff>=0.5",
   "types-requests>=2.32",
   "types-python-dateutil>=2.8"
 ]

[tool.setuptools.packages.find]
include = ["arbitrage_bot*"]

